Refactoring AutoFire settings page. [ci skip] #154 - Change FormControl componentClass propType from element to string; - Extract credentials from auth object in SettingsContainer and pass as prop; - Update AutoFire settings page route component on Widgets module route file; - Rename AutoFireForm file to AutoFireFormPage; - Add AutoFireFormPage into index file of Widget module settings directory.
